requirements.txt
gensim==4.3.3
h5py==3.13.0
kagglehub==0.3.11
matplotlib==3.10.1
numpy==2.2.4
pandarallel==1.6.5
pandas==2.2.3
pyarrow==19.0.1
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.2
spacy==3.8.4
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
